+++ static/views/admin/settings.php
@@ -78,9 +78,7 @@ if (get_option('permalink_structure') == '') {
     </div>
   <?php endif; ?>
 
-  <form id="wpp_settings_form" method="post"
-        action="<?php echo admin_url('edit.php?post_type=property&page=property_settings'); ?>"
-        enctype="multipart/form-data">
+  <form id="wpp_settings_form" method="post" action="<?php echo admin_url('edit.php?post_type=property&page=property_settings'); ?>" enctype="multipart/form-data">
     <?php wp_nonce_field('wpp_setting_save'); ?>
 
     <div id="wpp_settings_tabs" class="wpp_tabs clearfix">
@@ -114,14 +112,15 @@ if (get_option('permalink_structure') == '') {
             <th><?php _e('Options', ud_get_wp_property()->domain); ?></th>
             <td>
               <ul>
-                <li
-                  class="configuration_enable_comments"><?php echo WPP_F::checkbox("name=wpp_settings[configuration][enable_comments]&label=" . __('Enable comments', ud_get_wp_property()->domain), (isset($wp_properties['configuration']['enable_comments']) ? $wp_properties['configuration']['enable_comments'] : false)); ?></li>
-                <li class="configuration_enable_revisions"
-                    data-feature-since="2.0.0"><?php echo WPP_F::checkbox("name=wpp_settings[configuration][enable_revisions]&label=" . __('Enable revisions', ud_get_wp_property()->domain), (isset($wp_properties['configuration']['enable_revisions']) ? $wp_properties['configuration']['enable_revisions'] : false)); ?></li>
-                <li class="configuration_enable_layouts"
-                    data-feature-since="2.2.1"><?php echo WPP_F::checkbox("name=wpp_settings[configuration][enable_layouts]&label=" . __('Disable layouts', ud_get_wp_property()->domain), (isset($wp_properties['configuration']['enable_layouts']) ? $wp_properties['configuration']['enable_layouts'] : false)); ?></li>
-                <li
-                  class="wpp-setting-exclude-from-regular-search-results"><?php echo WPP_F::checkbox("name=wpp_settings[configuration][exclude_from_regular_search_results]&label=" . sprintf(__('Exclude %1s from regular search results.', ud_get_wp_property()->domain), $object_label['plural']), (isset($wp_properties['configuration']['exclude_from_regular_search_results']) ? $wp_properties['configuration']['exclude_from_regular_search_results'] : false)); ?></li>
+                <li class="configuration_enable_comments"><?php echo WPP_F::checkbox("name=wpp_settings[configuration][enable_comments]&label=" . __('Enable comments', ud_get_wp_property()->domain), (isset($wp_properties['configuration']['enable_comments']) ? $wp_properties['configuration']['enable_comments'] : false)); ?></li>
+                <li class="configuration_enable_revisions" data-feature-since="2.0.0"><?php echo WPP_F::checkbox("name=wpp_settings[configuration][enable_revisions]&label=" . __('Enable revisions', ud_get_wp_property()->domain), (isset($wp_properties['configuration']['enable_revisions']) ? $wp_properties['configuration']['enable_revisions'] : false)); ?></li>
+
+                <?php if( defined( 'WP_PROPERTY_LAYOUTS' ) && WP_PROPERTY_LAYOUTS ) { ?>
+                <li class="configuration_enable_layouts" data-feature-since="2.2.1"><?php echo WPP_F::checkbox("name=wpp_settings[configuration][enable_layouts]&label=" . __('Disable layouts', ud_get_wp_property()->domain), (isset($wp_properties['configuration']['enable_layouts']) ? $wp_properties['configuration']['enable_layouts'] : false)); ?></li>
+                <?php } ?>
+
+                <li class="wpp-setting-exclude-from-regular-search-results"><?php echo WPP_F::checkbox("name=wpp_settings[configuration][exclude_from_regular_search_results]&label=" . sprintf(__('Exclude %1s from regular search results.', ud_get_wp_property()->domain), $object_label['plural']), (isset($wp_properties['configuration']['exclude_from_regular_search_results']) ? $wp_properties['configuration']['exclude_from_regular_search_results'] : false)); ?></li>
+
               </ul>
             </td>
           </tr>
@@ -166,9 +165,9 @@ if (get_option('permalink_structure') == '') {
             </td>
           </tr>
 
-          <?php /* begin : Single Property Template Options */
-          //    echo $wp_properties[ 'configuration' ][ 'enable_layouts' ]; die;
-          if (!isset($wp_properties['configuration']['enable_layouts']) || $wp_properties['configuration']['enable_layouts'] != "false") : ?>
+          <?php if(defined( 'WP_PROPERTY_LAYOUTS' ) && WP_PROPERTY_LAYOUTS === false )  {
+
+            if( !isset($wp_properties['configuration']['enable_layouts']) || $wp_properties['configuration']['enable_layouts'] != "false")  { ?>
 
             <tr class="wpp-setting wpp-setting-single-template">
               <th><?php printf(__('Single %s Template', ud_get_wp_property()->domain), WPP_F::property_label()); ?></th>
@@ -242,8 +241,7 @@ if (get_option('permalink_structure') == '') {
                 </ul>
               </td>
             </tr>
-          <?php endif; ?>
-          <?php /* End : Single Property Template Options */ ?>
+          <?php } ?>
 
           <?php if ((!isset($wp_properties['configuration']['do_not_register_sidebars']) || (isset($wp_properties['configuration']['do_not_register_sidebars']) && $wp_properties['configuration']['do_not_register_sidebars'] != 'true')) && ($wp_properties['configuration']['enable_layouts'] == 'true')) : ?>
             <tr class="wpp-setting wpp-setting-widget-sidebars">
@@ -263,6 +261,8 @@ if (get_option('permalink_structure') == '') {
               </td>
             </tr>
           <?php endif; ?>
+
+          <?php } ?>
 
           <tr>
             <th><?php printf(__('Automatic Geolocation', ud_get_wp_property()->domain), WPP_F::property_label()); ?></th>
@@ -354,10 +354,8 @@ if (get_option('permalink_structure') == '') {
             </td>
           </tr>
 
-
           <?php do_action('wpp_settings_main_tab_bottom', $wp_properties); ?>
         </table>
-
 
       </div>